+++ packages/terra-clinical-result/tests/wdio/clinical-result-blood-pressure-spec.js
@@ -55,11 +55,11 @@ Terra.describeViewports('ClinicalResultBloodPressure', ['medium'], () => {
     Terra.it.validatesElement();
   });
 
-  describe('single results', () => {
-  before(() => {
-    browser.url('/#/raw/tests/terra-clinical-result/clinical-result/clinical-result-blood-pressure/single-result-clinical-result-blood-pressure');
-  });
+  describe('no data', () => {
+    before(() => {
+      browser.url('/#/raw/tests/terra-clinical-result/clinical-result/clinical-result-blood-pressure/no-data-clinical-result-blood-pressure');
+    });
 
-  Terra.it.validatesElement();
-});
+    Terra.it.validatesElement();
+  });
 });