+++ packages/terra-clinical-result/CHANGELOG.md
@@ -4,7 +4,7 @@ ChangeLog
 Unreleased
 -----------------
 ### Fixed
-* Fix bug when ClinicalBloodPresure was provided both a systolic and diastolic no data, presented error
+* Fixed bug where ClinicalBloodPresure if provided both a systolic and diastolic object with `resultNoData` true, presented error result instead of no data result.
 
 1.0.0 - (March 6, 2020)
 ----------